Resolved conflicts by combining changes from both branches:
- Added specta feature to altair-db dependency (from main)
- Kept altair-storage dependency (from storage branch)
- Combined storage commands with type exports in lib.rs
- Preserved all type exports (.typ<>() calls) from main
- Kept all storage-related commands from storage branch

Fixed type compatibility issue:
- Changed size_bytes from i64 to i32 in attachment queries
- Updated cast in storage commands to match

Conflicts resolved in:
- apps/guidance/src-tauri/Cargo.toml
- apps/guidance/src-tauri/src/lib.rs

+++ backend/crates/altair-commands/src/storage.rs
@@ -254,7 +254,7 @@ pub async fn storage_confirm_upload<C: surrealdb::Connection>(
         db,
         confirmation.filename.clone(),
         confirmation.mime_type.clone(),
-        confirmation.size_bytes as i64,
+        confirmation.size_bytes as i32,
         confirmation.storage_key.clone(),
         confirmation.checksum.clone(),
         db_media_type,
